refactor(types): split types.py into package `types/`
