+++ src/amrita_core/types/embedding.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from collections.abc import Sequence
+
+from amrita_core.types.base import BaseModel
+
+
+class EmbeddingChunk(BaseModel):
+    """Represents an embedding vector returned by embedding adapter."""
+
+    embedding: Sequence[float]
+
+    index: int
