+++ src/amrita_core/types/tool.py
@@ -1,0 +1,38 @@
+from __future__ import annotations
+
+import typing
+
+from pydantic import Field
+
+from amrita_core.types.base import BaseModel
+
+
+class Function(BaseModel):
+    arguments: str
+    """
+    The arguments to call the function with, as generated by the model in JSON
+    format. Note that the model does not always generate valid JSON, and may
+    hallucinate parameters not defined by your function schema. Validate the
+    arguments in your code before calling your function.
+    """
+
+    name: str
+    """The name of the function to call."""
+
+
+class ToolCall(BaseModel):
+    id: str
+    """The ID of the tool call."""
+
+    function: Function
+    """The function that the model called."""
+
+    type: typing.Literal["function"] = "function"
+    """The type of the tool. Currently, only `function` is supported."""
+
+
+class ToolResult(BaseModel):
+    role: typing.Literal["tool"] = Field(..., description="Role")
+    name: str = Field(..., description="Tool name")
+    content: str = Field(..., description="Tool return content")
+    tool_call_id: str = Field(..., description="Tool call ID")
